Create task 3, add class and methods, add tests (#3)
* Create task 3, add class and methods, add tests

* Fixes in test, and FileDatabase structure

* Add changes for reading file

---------

+++ app/src/main/java/pl/futurecollars/invoicing/db/Database.java
@@ -6,14 +6,14 @@ import pl.futurecollars.invoicing.model.Invoice;
 
 public interface Database {
 
-  int save(Invoice invoice);
+  long save(Invoice invoice);
 
-  Optional<Invoice> getById(int id);
+  Optional<Invoice> getById(long id);
 
   List<Invoice> getAll();
 
-  void update(int id, Invoice updatedInvoice);
+  void update(long id, Invoice updatedInvoice);
 
-  void delete(int id);
+  void delete(long id);
 
 }
